fix: deduplicate MCP results, log MCP injections to stats, honest learn_from_session message
- retrieval: deduplicate by block_id in _resolve_conflicts — same block
  appearing in multiple source indexes was returned twice (seen in Dino's
  Codex session with nextjs15-approuter-build-version-16-0-0)
- mcp_server: list_suggested_blocks now writes to hook_log.jsonl so
  get_stats reflects MCP-first users (was always 0 for Codex users)
- mcp_server: learn_from_session return message no longer claims a daemon
  will process it — directs user to `turnzero harvest` instead
- tests: 4 new tests covering dedup, injection logging, and honest message
- standards: bring ruff + mypy to clean across all source files (37 → 0
  ruff errors, 14 → 0 mypy errors); add coding/git/branching standards
  to CLAUDE.md

+++ turnzero/analytics.py
@@ -25,7 +25,7 @@ class SessionAnalytics:
     # Rough estimates — 1 prior ≈ 1 avoided correction turn, ~1500 tokens, ~4 min
     TURNS_SAVED_PER_PRIOR: float = 2.5
     TOKENS_PER_TURN: int = 1500
-    MINUTES_PER_TURN: int = 4.0
+    MINUTES_PER_TURN: float = 4.0
 
     def log_injection(self, block_ids: list[str]) -> None:
         self.events.append(SessionEvent(
